requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.12.0
notion-client==2.2.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
